Update module to use floats and added tests

+++ arithmetic/addition.go
@@ -1,7 +1,7 @@
 // arithmetic/addition.go
 package arithmetic
 
-// Add returns the sum of two integers.
-func Add(a, b int) int {
+// Add returns the sum of two float64 numbers.
+func Add(a, b float64) float64 {
     return a + b
 }
