+++ arithmetic/multiplication.go
@@ -1,7 +1,7 @@
 // arithmetic/multiplication.go
 package arithmetic
 
-// Multiply returns the product of two integers.
-func Multiply(a, b int) int {
-	return a * b
+// Multiply returns the product of two float64 numbers.
+func Multiply(a, b float64) float64 {
+    return a * b
 }
